refactor(ui): rename core folder
