+++ webapp/src/components/common/Matrix/components/MatrixGrid/MatrixGrid.test.tsx
@@ -17,10 +17,10 @@ import userEvent from "@testing-library/user-event";
 import Box from "@mui/material/Box";
 import MatrixGrid, { MatrixGridProps } from ".";
 import SplitView from "../../../SplitView";
-import type { EnhancedGridColumn } from "../../core/types";
+import type { EnhancedGridColumn } from "../../shared/types";
 import { mockGetBoundingClientRect } from "../../../../../tests/mocks/mockGetBoundingClientRect";
 import { mockHTMLCanvasElement } from "../../../../../tests/mocks/mockHTMLCanvasElement";
-import { Column } from "../../core/constants";
+import { Column } from "../../shared/constants";
 
 interface RenderMatrixOptions {
   width?: string;
