+++ webapp/src/components/common/Matrix/hooks/useGridCellContent/__tests__/useGridCellContent.test.ts
@@ -12,7 +12,7 @@
  * This file is part of the Antares project.
  */
 
-import { Column } from "../../../core/constants";
+import { Column } from "../../../shared/constants";
 import {
   createColumn,
   AGGREGATE_DATA,
