+++ webapp/src/components/common/Matrix/hooks/useGridCellContent/types.ts
@@ -13,7 +13,7 @@
  */
 
 import { GridCell, Item } from "@glideapps/glide-data-grid";
-import type { EnhancedGridColumn, MatrixAggregates } from "../../core/types";
+import type { EnhancedGridColumn, MatrixAggregates } from "../../shared/types";
 
 export type GridToDataFunction = (cell: Item) => Item | null;
